+++ app/services/artwork/version/generator/build.layers.service.ts
@@ -4,7 +4,7 @@ import {
 } from '#app/definitions/artwork-generator'
 import { type IArtworkVersionWithChildren } from '#app/models/artwork-version/artwork-version.server'
 import { filterAssetsVisible, groupAssetsByType } from '#app/models/asset/utils'
-import { findManyDesignsWithType } from '#app/models/design/design.server'
+import { findManyDesignsWithType } from '#app/models/design/design.get.server'
 import { getArtworkVersionVisiblePalettes } from '#app/models/design-artwork-version/design-artwork-version.server'
 import { getLayerVisiblePalettes } from '#app/models/design-layer/design-layer.server'
 import { type ILayerWithChildren } from '#app/models/layer/layer.server'
