+++ app/utils/prisma-extensions-artwork-version.ts
@@ -1,5 +1,4 @@
 import { Prisma } from '@prisma/client'
-import { type whereArgsType } from '#app/schema/artwork'
 import { prismaExtended } from './db.server'
 
 // must be in /utils to actually connect to prisma
@@ -49,7 +48,7 @@ export type ExtendedArtworkVersion = Prisma.Result<
 export const findFirstArtworkVersionInstance = async ({
 	where,
 }: {
-	where: whereArgsType
+	where: { id: string }
 }): Promise<ExtendedArtworkVersion | null> => {
 	return await prismaExtended.artworkVersion.findFirst({
 		where,
