+++ app/utils/prisma-extensions-artwork.ts
@@ -1,5 +1,4 @@
 import { Prisma } from '@prisma/client'
-import { type whereArgsType } from '#app/schema/artwork'
 import { prismaExtended } from './db.server'
 
 // must be in /utils to actually connect to prisma
@@ -49,7 +48,7 @@ export type ExtendedArtwork = Prisma.Result<
 export const findFirstArtworkInstance = async ({
 	where,
 }: {
-	where: whereArgsType
+	where: { id: string }
 }): Promise<ExtendedArtwork | null> => {
 	return await prismaExtended.artwork.findFirst({
 		where,
