+++ app/utils/prisma-extensions-rotate.ts
@@ -1,5 +1,4 @@
 import { Prisma } from '@prisma/client'
-import { type whereArgsType } from '#app/schema/design'
 import { prismaExtended } from './db.server'
 
 // must be in /utils to actually connect to prisma
@@ -49,7 +48,7 @@ export type ExtendedRotate = Prisma.Result<
 export const findFirstRotateInstance = async ({
 	where,
 }: {
-	where: whereArgsType
+	where: { id: string }
 }): Promise<ExtendedRotate | null> => {
 	return await prismaExtended.rotate.findFirst({
 		where,
